Vision/yolo-ann-snn-batch/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
_pytorch_select=0.2=gpu_0
blas=1.0=mkl
ca-certificates=2022.5.18.1=ha878542_0
cffi=1.15.0=py37h036bc23_0
colorama=0.4.4=pyh9f0ad1d_0
cudatoolkit=9.2.148=h60dc4a4_10
cudnn=7.6.5.32=h5754881_1
cycler=0.11.0=pyhd8ed1ab_0
freetype=2.10.4=h0708190_1
giflib=5.2.1=h36c2ea0_2
icu=67.1=he1b5a44_0
importlib-metadata=4.11.4=py37h89c1867_0
jpeg=9e=h166bdaf_1
kiwisolver=1.4.2=py37h7cecad7_1
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libdeflate=1.10=h7f98852_0
libffi=3.4.2=h7f98852_5
libgcc-ng=12.1.0=h8d9b700_16
libnsl=2.0.0=h7f98852_0
libpng=1.6.37=h21135ba_2
libprotobuf=3.20.1=h6239696_0
libstdcxx-ng=12.1.0=ha89aaad_16
libtiff=4.4.0=h0fcbabc_0
libwebp=1.2.2=h3452ae3_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libzlib=1.2.12=h166bdaf_0
llvm-openmp=14.0.4=he0ac6c6_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.7=pyhd8ed1ab_0
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py37h1d35a4c_1
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py37h8f50634_2
mkl_fft=1.3.0=py37h902c9e0_1
mkl_random=1.2.0=py37h9fdb41a_1
ncurses=6.3=h27087fc_1
ninja=1.7.2=0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.6.0.66=pypi_0
openjpeg=2.4.0=hb52868f_1
openssl=3.0.3=h166bdaf_0
pillow=6.1.0=py37h34e0f95_0
pip=22.1.2=pyhd8ed1ab_0
protobuf=3.20.1=py37hd23a5d3_0
pthread-stubs=0.4=h36c2ea0_1001
pycparser=2.21=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
python=3.7.12=hf930737_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
readline=8.1.2=h0f457ee_0
setuptools=62.3.4=py37h89c1867_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.38.5=h4ff8645_0
tensorboard=1.14.0=py37_0
tk=8.6.12=h27826a3_0
torch=1.4.0=pypi_0
torchvision=0.5.0=pypi_0
tornado=6.1=py37h540881e_3
tqdm=4.64.0=pyhd8ed1ab_0
typing-extensions=4.2.0=hd8ed1ab_1
typing_extensions=4.2.0=pyha770c72_1
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.37.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=h166bdaf_0
zstd=1.5.2=h8a70e8d_1
